链接crud前后端联调成功，差前端体验问题需要完善；

+++ src/Dotnet9.ASPNETCore/ResponseResults/ResponseResult.cs
@@ -7,7 +7,7 @@ public class ResponseResult<T>
 
     public string? ErrorCode
     {
-        get { return Code.ToString(); }
+        get { return ((int)Code).ToString(); }
     }
 
     public string? ErrorMessage { get; set; }
